Add multi folder permission flow
